requirements.txt
carla==0.9.15
matplotlib>=3.8.0
Pillow>=10.0.0
